Updated the GeneratePEMCert interface
Extensions for the Certificates now supported so we can specify that these
certificates can be used as a Certificate Authority, to sign certificates,
to be used for client auth, and to be used for server auth. To use for a TLS
server you can specify domain names and/or IP addresses.

Also tossed in an early version of the TLS demo:

  genkeys.go: Generate the JWKs for the client and server.
  server.go: Set the client and server keys to use and it runs a simple
    HTTPS server on localhost:8888
  client.go: Set the client and server keys to use and it will make a
    couple of requests to https://localhost:8888 and https://127.0.0.1:8888