Remove unnecessary use of pointer types
To simplify the code, some pointer type fields have been removed where it makes
sense. Doing so should reduce the number of scattered heap allocations and
simplify areas where we copy into a struct.

There is probably more work that can be done to get the benefits of this change.

Signed-off-by: Stephen J Day <stephen.day@docker.com>